generalize 'unifyTypes'
